% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parties_info.R
\name{.fetch_parties_info}
\alias{.fetch_parties_info}
\title{Fetch party information from the API (Internal)}
\usage{
.fetch_parties_info(states, inicio_vigencia, fim_vigencia)
}
\arguments{
\item{states}{A character vector with state abbreviations.}

\item{inicio_vigencia}{A string in "dd/mm/yyyy" format representing the start date.}

\item{fim_vigencia}{A string in "dd/mm/yyyy" format representing the end date.}
}
\value{
A dataframe with party information.
}
\description{
This function retrieves party information from the TSE API
based on the given states and validity dates.
}
\keyword{internal}
