% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_party_members.R
\name{.fetch_party_members}
\alias{.fetch_party_members}
\title{Fetch party members from API (Internal)}
\usage{
.fetch_party_members(party_id)
}
\arguments{
\item{party_id}{Character representing the party ID.}
}
\value{
A tibble with party members.
}
\description{
Retrieves party members based on the given party ID.
}
\keyword{internal}
