% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_party_members.R
\name{.parse_members}
\alias{.parse_members}
\title{Parse API response into tibble (Internal)}
\usage{
.parse_members(response, party_id)
}
\arguments{
\item{response}{List containing the API response.}

\item{party_id}{Character representing the party ID.}
}
\value{
A tibble containing the parsed party members.
}
\description{
Converts the API response into a structured tibble.
}
\keyword{internal}
