% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_party_members.R
\name{.request_party_data}
\alias{.request_party_data}
\title{Request party data from API (Internal)}
\usage{
.request_party_data(party_id)
}
\arguments{
\item{party_id}{Character representing the party ID.}
}
\value{
A list containing the API response.
}
\description{
Sends a request to the API to retrieve party data.
}
\keyword{internal}
