% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.validate_id_orgao_partidario}
\alias{.validate_id_orgao_partidario}
\title{Validate party organization ID (Internal)}
\usage{
.validate_id_orgao_partidario(id_orgao_partidario)
}
\arguments{
\item{id_orgao_partidario}{Character vector representing the party organization ID.}
}
\value{
Validated character vector of party organization IDs.
}
\description{
Ensures that the input is a character vector and checks for validity.
}
\keyword{internal}
