% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parties_info.R
\name{get_parties_info}
\alias{get_parties_info}
\title{Get the ID of all the Parties of the Brazil or a specific state}
\usage{
get_parties_info(states = NULL, ivigencia = NULL, fvigencia = NULL)
}
\arguments{
\item{states}{A character vector containing the abbreviation of the states of
Brazil or the Federal District. If nothing is inserted, all states will be
selected.}

\item{ivigencia}{A string that contains a date, must be in "dd/mm/yyyy" format,
that marks the beginning of the search. If nothing is inserted, the system date
will be used.}

\item{fvigencia}{A string that contains a date, must be in "dd/mm/yyyy" format,
that marks the end of the search. If nothing is inserted, the system date will
be used.}
}
\value{
A dataframe containing the ID and information of the Parties
}
\description{
Get the ID of all the Parties of the Brazil or a specific state
}
\examples{
\dontrun{

 parties_id <- get_party_id(
  states = c("AL", "AC", "AM"),
  ivigencia = "20/05/2017",
  fvigencia = "20/02/2022"
 )

}
}
