% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_party_id.R
\name{get_party_id}
\alias{get_party_id}
\title{Get the ID of all the Parties of the Brazil or a specific state}
\usage{
get_party_id(states = NULL)
}
\arguments{
\item{states}{A character vector containing the abbreviation of the states of
Brazil. If nothing is inputed then will select all states}
}
\value{
A dataframe containing the ID and information of the Parties
}
\description{
Get the ID of all the Parties of the Brazil or a specific state
}
\examples{
\dontrun{

 parties_id <- get_party_id(states = c("AL", "AC", "AM"))

}
}
