% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_party_members.R
\name{get_party_members}
\alias{get_party_members}
\title{Get party members from a specific party organization}
\usage{
get_party_members(id_orgao_partidario = NULL)
}
\arguments{
\item{id_orgao_partidario}{An integer representing the ID of the party organization.
If nothing is provided, it will run for all parties in Brazil at municipality level.}
}
\value{
A tibble containing the members of the party organization.
}
\description{
This function makes a request to the SGIP API to retrieve members of a party organization,
identified by \code{id_orgao_partidario}, and returns the data as a tibble.
}
\examples{
\dontrun{
  members <- get_party_members(12345)
  print(members)
}
}
